requirements.txt
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
click==7.1.2
colorama==0.4.4
Django==3.2
factory-boy==3.2.0
Faker==8.1.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.8.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest-django==4.2.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2